Mon, Jan 1, 2024, 2:54 AM -05:00

+++ filter-less/helpers.c
@@ -8,7 +8,7 @@ void grayscale(int height, int width, RGBTRIPLE image[height][width])
     {
         for (int j = 0; j < width; j++)
         {
-         int average = image[i][j].rgbtRed + image[i][j].rgbtGreen + image[i][j].rgbtBlue;
+         float average = image[i][j].rgbtRed + image[i][j].rgbtGreen + image[i][j].rgbtBlue;
          int averagePixel = round(average/3);
 
          image[i][j].rgbtRed   = averagePixel;
